pyproject.toml
[tool.poetry]
name = "sdmitriev"
version = "0.1.0"
description = ""
authors = ["Sergei Dmitriev <graywrk@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.2"
gunicorn = "^21.2.0"
pydantic-settings = "^2.1.0"
psycopg2-binary = "^2.9.9"
django-mdeditor = "^0.1.20"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
